# Simple left-anchored patterns load and match faster